+++ src/main/java/domain/service/BlackJackResultMaker.java
@@ -1,5 +1,9 @@
-package domain.model;
+package domain.service;
 
+import domain.model.Dealer;
+import domain.model.Player;
+import domain.model.Result;
+import domain.model.Score;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
